requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.0
django-cors-headers==4.6.0
django-jazzmin==3.0.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
phonenumbers==8.13.51
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
whitenoise==6.8.2